restructure unit test project and add TestHelpers class
